Embed the egm96 file
In order to find the offset between altitude above sea level and altitude above the WGS84 ellipsoid, we need the EGM96 data. This data describes the shape of Earth's gravitational field. And we use this to calculate the offset.

However, up till now we've been separately installing the EGM96 data. This has worked fine on debian-like distros because apt can install the files we need. But this didn't work on all other operating systems.

To address this we now embed the emg96 data in the whl package.

#3