Make EGM96 Data Easier to Use in Development Containers
Now, when you use this module in a development container, you don't need to manually embed the EGM96 data. The Dockerfile installs this data in /usr/share, and the module will use this path if it doesn't find the data embedded in the package.

However, when distributing the package, you still need to manually embed the EGM96 data. This update just makes development easier.